FEAT: Changed int to double for division, changed id to long for posts, used PostMapping instead of create post

+++ src/main/java/com/codeup/springproject/controllers/MathController.java
@@ -29,7 +29,7 @@ class MathController {
 
     @GetMapping("/divide/{num1}/by/{num2}")
     @ResponseBody
-    public int divide(@PathVariable int num1, @PathVariable int num2) {
+    public double divide(@PathVariable double num1, @PathVariable double num2) {
         return num1 / num2;
     }
